Split out `getFileUploader`
